Setting up the controllers for the H2 database belongs to that Spring boot project

+++ basicCRUD-H2/src/main/java/com/toadsdewin/basicCRUDH2/Services/PersonService.java
@@ -66,7 +66,7 @@ public class PersonService implements PersonServiceInterface
         return personInterface.findByEmail(email);
     }
     @Override
-    public ArrayList<Person> findByPhone(Integer phone) {
+    public ArrayList<Person> findByPhone(Long phone) {
         return personInterface.findByPhone(phone);
     }
     @Override
